+++ core-ng-mongo-test/src/main/java/core/framework/mongo/module/TestMongoConfig.java
@@ -7,26 +7,34 @@ import de.bwaldvogel.mongo.MongoServer;
 import de.bwaldvogel.mongo.backend.memory.MemoryBackend;
 
 import java.net.InetSocketAddress;
-import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * @author neo
  */
 public class TestMongoConfig extends MongoConfig {
-    private static final AtomicInteger NEXT_MONGO_SERVER_PORT = new AtomicInteger(27017);
-    int port;
+    private static MongoServer server;
+    String name;
 
     @Override
     protected void initialize(ModuleContext context, String name) {
         super.initialize(context, name);
-        var server = new MongoServer(new MemoryBackend());
-        port = NEXT_MONGO_SERVER_PORT.getAndIncrement();
-        server.bind(new InetSocketAddress("localhost", port));
-        context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> server.shutdown());
+        this.name = name;
+
+        startInMemoryMongoServer(context);
+    }
+
+    private void startInMemoryMongoServer(ModuleContext context) {
+        // in test env, config is initialized in order and within same thread, so no threading issue
+        if (server == null) {
+            server = new MongoServer(new MemoryBackend());
+            server.bind(new InetSocketAddress("localhost", 27017));
+            context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> server.shutdown());
+        }
     }
 
     @Override
     ConnectionString connectionString(ConnectionString uri) {
-        return new ConnectionString("mongodb://localhost:" + port + "/test");
+        String database = name == null ? "test" : name;
+        return new ConnectionString("mongodb://localhost:27017/" + database);
     }
 }