+++ core-ng-search-test/src/main/java/core/framework/search/module/TestSearchConfig.java
@@ -9,12 +9,31 @@ import core.framework.search.impl.LocalElasticSearch;
  * @author neo
  */
 public class TestSearchConfig extends SearchConfig {
+    private static LocalElasticSearch server;
+
+    static {
+        // refer to io.netty.util.NettyRuntime.AvailableProcessorsHolder.setAvailableProcessors
+        // refer to org.elasticsearch.transport.netty4.Netty4Utils.setAvailableProcessors
+        // netty only allows set available processors once
+        System.setProperty("es.set.netty.runtime.available.processors", "false");
+    }
+
     @Override
     protected void initialize(ModuleContext context, String name) {
         super.initialize(context, name);
-        var search = new LocalElasticSearch();
-        search.start();
-        context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> search.close());
+
+        startLocalElasticSearch(context);
+    }
+
+    // currently multiple elasticsearch in one service will share one test local elasticsearch, therefor there might be document type collation (e.g. same document type on both es)
+    // at this point this behavior is not causing big problem (can be workaround in unit test), so we keep this simple rather than start multiple local es on different port which is more complex and slow
+    private void startLocalElasticSearch(ModuleContext context) {
+        // in test env, config is initialized in order and within same thread, so no threading issue
+        if (server == null) {
+            server = new LocalElasticSearch();
+            server.start();
+            context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> server.close());
+        }
     }
 
     @Override